+++ NEWS.md
@@ -1,5 +1,11 @@
 # greta (development version)
 
+# greta 0.4.2.9000
+
+## Fixes
+
+* Issue where `future` and `parallely` packages error when a CPU with only one core is provided (#513, #516).
+
 # greta 0.4.2
 
 ## Fixes